APIREST

+++ templates_c/141270abe502594041f6f7cb34892a050f07987b_0.file.render.array.pelis.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 4.3.4, created on 2023-10-13 20:52:43
+/* Smarty version 4.3.4, created on 2023-11-12 21:39:34
   from 'C:\xampp\htdocs\trabajotp1\templates\render.array.pelis.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.3.4',
-  'unifunc' => 'content_652991fb6ea553_53430863',
+  'unifunc' => 'content_65513806e7fee0_38863217',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     '141270abe502594041f6f7cb34892a050f07987b' => 
     array (
       0 => 'C:\\xampp\\htdocs\\trabajotp1\\templates\\render.array.pelis.tpl',
-      1 => 1697148138,
+      1 => 1697487267,
       2 => 'file',
     ),
   ),
@@ -20,7 +20,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   array (
   ),
 ),false)) {
-function content_652991fb6ea553_53430863 (Smarty_Internal_Template $_smarty_tpl) {
+function content_65513806e7fee0_38863217 (Smarty_Internal_Template $_smarty_tpl) {
 ?>
     <h1><?php echo $_smarty_tpl->tpl_vars['titulo']->value;?>
 </h1>
@@ -38,10 +38,12 @@ $_smarty_tpl->tpl_vars['peli']->do_else = false;
 " alt="">
     </div>
     <div class="info-peliculas">
+    <?php if ($_smarty_tpl->tpl_vars['user']->value == 1) {?>
       <a href="eliminar/<?php echo $_smarty_tpl->tpl_vars['peli']->value->id;?>
 " class="nav-link"><button>Borrar</button></a> 
       <a href="modificarview/<?php echo $_smarty_tpl->tpl_vars['peli']->value->id;?>
 " class="nav-link"><button>Modificar</button></a> 
+    <?php }?>
       <h1 class="home-titulo-pelicula"><a href="pelicula/<?php echo $_smarty_tpl->tpl_vars['peli']->value->id;?>
 "><?php echo $_smarty_tpl->tpl_vars['peli']->value->name;?>
 </a></h1>
@@ -54,22 +56,23 @@ $_smarty_tpl->tpl_vars['peli']->do_else = false;
 <?php
 }
 $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?> 
-
-<div>
-        <h2 class="">Agregar un pelicula a la lista!</h2>
-        <form method="post" action="createPeli" enctype="multipart/form-data">
-            <label>Nombre de la Pelicula:</label><input type="text" name="name" required>
-            <label>Categoria de la pelicula:</label><select name="categorie">
-                <option value="action">Accion</option>
-                <option value="adventure">Aventura</option>
-                <option value="drama">Drama</option>
-            </select>
-            <label>Año de lanzamiento:</label><input type="text" name="release_date" required>
-            <label>Descripcion:</label><input type="text" name="description" required>
-            <label>Adjuntar imagen:</label><input type="file" name="img">
-            <input type="submit" value="Agregar">
-        </form>
-</div>
+<?php if ($_smarty_tpl->tpl_vars['user']->value == 1) {?>
+  <div class="contenedor-agregar-pelicula">
+    <h2 class="">Agregar un pelicula a la lista!</h2>
+    <form class="formulario-agregar-peli" method="post" action="createPeli" enctype="multipart/form-data">
+      <label>Nombre de la Pelicula:</label><input type="text" name="name" required>
+      <label>Categoria de la pelicula:</label><select name="categorie">
+          <option value="action">Accion</option>
+          <option value="adventure">Aventura</option>
+          <option value="drama">Drama</option>
+      </select>
+      <label>Año de lanzamiento:</label><input type="text" name="release_date" required>
+      <label>Descripcion:</label><input type="text" name="description" required>
+      <label>Adjuntar imagen resolucion estandar 300x445:</label><input type="file" name="img">
+      <input type="submit" value="Agregar">
+    </form>
+  </div>
+<?php }?>
 
             <?php }
 }
